LayoutTests/canvas/philip/tests/2d.line.miter.lineedge.html

<!DOCTYPE html>
<title>Canvas test: 2d.line.miter.lineedge</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.line.miter.rightangle.html" title="2d.line.miter.rightangle">
<link rel="next" href="2d.line.miter.within.html" title="2d.line.miter.within">
<body class="show_output">
<p>
 <a href="2d.line.miter.rightangle.html" accesskey="p" title="[p] 2d.line.miter.rightangle">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.line.miter.within.html" accesskey="n" title="[n] 2d.line.miter.within">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.miter.html">miter</a>.lineedge</h1>
<p class="desc">Miter joins are not drawn when the miter limit is exceeded at the corners of a zero-height rectangle</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.lineJoin.miter">2d.lineJoin.miter</a>
<li><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=401791">Bugzilla</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.lineWidth = 200;
ctx.lineJoin = 'miter';

ctx.strokeStyle = '#f00';
ctx.miterLimit = 1.414;
ctx.beginPath();
ctx.strokeRect(100, 25, 200, 0);

_assertPixel(canvas, 1,1, 0,255,0,255, "1,1", "0,255,0,255");
_assertPixel(canvas, 48,1, 0,255,0,255, "48,1", "0,255,0,255");
_assertPixel(canvas, 48,48, 0,255,0,255, "48,48", "0,255,0,255");
_assertPixel(canvas, 1,48, 0,255,0,255, "1,48", "0,255,0,255");


});
</script>
