LayoutTests/canvas/philip/tests/2d.line.miter.lineedge-expected.txt

< [index] >
2d.line.miter.lineedge
Miter joins are not drawn when the miter limit is exceeded at the corners of a zero-height rectangle
References: 2d.lineJoin.miter Bugzilla
Actual output:
Expected output:

Failed assertion: got pixel [255,0,0,255] at (1,1), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (48,1), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (48,48), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (1,48), expected [0,255,0,255]
